+++ spotify_videos/__main__.py
@@ -125,28 +125,10 @@ def play_videos_dbus(player: Union['VLCPlayer', 'MpvPlayer'],
         spotify.wait()
 
 
-def play_videos_windows(player: Union['VLCPlayer', 'MpvPlayer'],
-                        spotify: 'WindowsAPI') -> None:
+def play_videos_swspotify(player: Union['VLCPlayer', 'MpvPlayer'],
+                        spotify: 'SwSpotifyAPI') -> None:
     """
-    Playing videos with the Windows "API" from SwSpotify.
-
-    It's really basic and can only get the window title.
-    """
-
-    while True:
-        url = get_url(spotify.artist, spotify.title)
-        player.start_video(url)
-
-        if config.lyrics:
-            print_lyrics(spotify.artist, spotify.title)
-
-        spotify.wait()
-
-
-def play_videos_darwin(player: Union['VLCPlayer', 'MpvPlayer'],
-                      spotify: 'DarwinAPI') -> None:
-    """
-    Playing videos with the Darwin (macOS) "API" from SwSpotify.
+    Playing videos with the SwSpotify API.
 
     It's really basic and can only get the window title.
     """
@@ -233,16 +215,11 @@ def choose_platform() -> None:
         msg = "Waiting for a Spotify session to be ready..."
         if wait_for_connection(dbus_spotify.connect, msg):
             play_videos_dbus(dbus_spotify.player, dbus_spotify)
-    elif system == "Windows" and not config.use_web_api:
-        from .api.windows import WindowsAPI
-        windows_api = WindowsAPI(logger)
+    elif system in ('Windows', 'Darwin') and not config.use_web_api:
+        from .api.swspotify import SwSpotifyAPI
+        swspotify_api = SwSpotifyAPI(logger)
 
-        play_videos_windows(player, windows_api)
-    elif system == "Darwin" and not config.use_web_api:
-        from .api.darwin import DarwinAPI
-        darwin_api = DarwinAPI(logger)
-
-        play_videos_darwin(player, darwin_api)
+        play_videos_swspotify(player, swspotify_api)
     else:
         from .api.web import WebAPI
         web_spotify = WebAPI(player, logger, config.client_id,